done the refactor of whole page
