+++ src/app/(public)/about/components/AboutBanner.tsx
