+++ src/app/(public)/about/components/InfoBody.tsx
