+++ src/app/(public)/about/components/TeamCard.tsx
