+++ src/app/(public)/blogs/detail-blog/components/HeroSection.tsx
