+++ src/app/(public)/blogs/detail-blog/layout.tsx
