+++ src/app/(public)/features/components/FeaturesSection.tsx
