+++ src/app/(public)/landing/components/FeaturesSection.tsx
