+++ src/app/(public)/landing/components/HeroSection.tsx
