+++ src/app/(public)/landing/page.tsx
@@ -1,5 +1,5 @@
-import HeroSection from '@/app/landing/components/HeroSection';
-import TestimonialsSection from '@/app/landing/components/TestimonialsSection';
+import HeroSection from '@/app/(public)/landing/components/HeroSection';
+import TestimonialsSection from '@/app/(public)/landing/components/TestimonialsSection';
 
 import FeaturesSection from './components/FeaturesSection';
 import PlanSection from './components/PlanSection/PlanSection';
