+++ src/app/(public)/pricing/components/PricingSection.tsx
