+++ src/app/(public)/signin/page.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import styled from 'styled-components';
 
-import SigninForm from '@/app/signin/component/SigninForm';
+import SigninForm from '@/app/(public)/signin/component/SigninForm';
 
 const PageContainer = styled.div`
   min-height: 100vh;
