+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Landing from '@/app/landing/page';
+import Landing from '@/app/(public)/landing/page';
 import MainLayout from '@/components/layout/main-layout';
 
 export default function LandingPage() {
